feat: Change/addition of files due to the addition of vite
