--- work16/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet"
-        href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome-animation/0.0.10/font-awesome-animation.css"
-        type="text/css" media="all" />
-    <link rel="stylesheet" href="./style.css" type="text/css" />
-    <script defer src="https://use.fontawesome.com/releases/v5.0.6/js/all.js"></script>
-    <script defer src="https://use.fontawesome.com/releases/v5.1.0/js/all.js"
-        integrity="sha384-3LK/3kTpDE/Pkp8gTNp2gR/2gOiwQ6QaO7Td0zV76UFJVhqLl4Vl3KL1We6q6wR9"
-        crossorigin="anonymous"></script>
-    <script src=" ./index.js" defer></script>
-    <title>TerraceTechフロントエンドエンジニア養成所</title>
-</head>
-
-<body>
-    <div class="newsContent">
-        <ul role="tablist" class="tabList" id="js_tabList"></ul>
-    </div>
-</body>
-
-</html>
