--- work16/index.js
@@ -1,258 +1,0 @@
-
-// import { format, formatDistance, formatRelative, subDays } from 'date-fns';
-const REQUEST_URL = "https://mocki.io/v1/e611647b-bcef-4351-81cd-118512c7690a"
-
-// 間違っているURLの場合↓
-// const REQUEST_URL = "https://myjson.dit.upm.es/api/bins/ほげほげajy3";
-// 空の配列の場合↓
-// const REQUEST_URL = "https://mocki.io/v1/242b685f-a3d7-45a8-aeca-0376bd495b89";
-// 503エラーの場合↓
-// const REQUEST_URL = "https://httpstat.us/503";
-
-const request = async () => {
-    const response = await fetch(REQUEST_URL, {
-        headers: {
-            "Content-Type": "application/json"
-        }
-    });
-    if (!response.ok) {
-        throw new Error(`${response.status}:${response.statusText}`);
-    } else {
-        return response.json();
-    }
-}
-
-const init = async () => {
-    renderLoadingImg();
-
-    let responseJsonData;
-
-    try {
-        responseJsonData = await request();
-
-        if (!responseJsonData) {
-            return;
-        }
-    } catch (error) {
-        displayErrorMessage(error.message);
-    }
-    finally {
-        hideLoadingImg();
-    }
-
-    if (!responseJsonData.data.length) {
-        displayErrorMessage("data is empty");
-        return;
-    }
-    displayNews(responseJsonData.data);
-}
-
-const displayErrorMessage = (error) => {
-    const p = document.createElement("p");
-    p.textContent = error;
-    document.body.prepend(p);
-}
-
-const displayNews = (newsDataArray) => {
-    renderNewsTab(newsDataArray);
-    renderNewsContent(newsDataArray);
-}
-
-const renderNewsTab = (newsDataArray) => {
-    const tabList = document.getElementById("js_tabList");
-    const fragment = document.createDocumentFragment();
-
-    newsDataArray.forEach((newsCategoryObj, newsCategoryIndex) => {
-        const li = document.createElement("li");
-        const a = document.createElement("a");
-        let isInitialDisplay = newsCategoryObj.isInitialDisplay;
-
-        if (isInitialDisplay === "true") {
-            li.setAttribute("aria-selected", true);
-        } else {
-            li.setAttribute("aria-selected", false);
-        }
-        li.setAttribute("roll", "tab");
-        li.id = `${"js_tabTopics" + ++newsCategoryIndex}`;
-        li.classList = "tabTopics";
-
-        a.textContent = newsCategoryObj.field
-        a.href = "#"
-
-        li.appendChild(a);
-        fragment.appendChild(li);
-    });
-    tabList.appendChild(fragment);
-
-    clickEventChangeTabAttribute();
-}
-
-const renderNewsContent = (newsDataArray) => {
-    newsDataArray.forEach((newsCategoryObj, newsCategoryIndex) => {
-        const tabTopics = document.querySelectorAll(".tabTopics");
-        const newsContentsData = newsCategoryObj.contents;
-        const newsCategoryImg = newsCategoryObj.img;
-        const newsArticleDate = newsCategoryObj.date;
-        let isInitialDisplay = newsCategoryObj.isInitialDisplay;
-
-        console.log(newsArticleDate);
-
-        if (isInitialDisplay === "true") {
-            const div = document.createElement("div");
-            const newsSection = createSection(tabTopics, newsCategoryIndex);
-
-            div.classList = "tabpanelTopics_wrap";
-
-            document.getElementById("js_tabList").after(newsSection);
-            div.appendChild(createTopicImg(newsCategoryImg));
-            div.prepend(renderArticle(newsContentsData, newsCategoryIndex));
-
-            newsSection.appendChild(div);
-        }
-
-        tabTopics[newsCategoryIndex].addEventListener("click", (e) => {
-            removeCommentIcon();
-
-            newsContentsData.forEach((newsArticleData, categoryNewsArticleDataIndex) => {
-                const commentArray = newsArticleData.comments;
-
-                changeCategoryConetent(newsCategoryIndex, clickedTabElement = e.currentTarget);
-                changeCategoryTitle(newsArticleData, categoryNewsArticleDataIndex);
-                changeCategoryImg(newsCategoryImg);
-
-                if (commentArray.length !== 0) {
-                    const liElements = document.querySelectorAll(".tabpanelTopics_wrap li");
-                    const liElement = liElements[categoryNewsArticleDataIndex];
-
-                    liElement.append(createCommentIcon(commentArray));
-                }
-            })
-        })
-    })
-}
-
-const isLatestArticles = () => {
-
-}
-
-const clickEventChangeTabAttribute = () => {
-    const tabTopics = document.querySelectorAll(".tabTopics");
-    for (let i = 0; i < tabTopics.length; i++) {
-        tabTopics[i].addEventListener("click", (event) => {
-            const selectedTab = document.querySelector('[aria-selected="true"]');
-            selectedTab.ariaSelected = false;
-            event.currentTarget.ariaSelected = true;
-        });
-    }
-}
-
-const removeCommentIcon = () => {
-    const commentIconWrapElements = document.querySelectorAll(".commentIcon_wrap");
-    for (let k = 0; k < commentIconWrapElements.length; k++) {
-        commentIconWrapElements[k].remove();
-    }
-}
-
-const changeCategoryTitle = (newsArticleData, categoryNewsArticleDataIndex) => {
-    const currentNewsContentAncorElements = document.querySelectorAll(`section[roll="tabpanel"] a`);
-    currentNewsContentAncorElements[categoryNewsArticleDataIndex].textContent = newsArticleData.title;
-}
-
-const changeCategoryImg = (newsCategoryImg) => {
-    const currentNewsContentImgElem = document.querySelector(".tabTopicImg > img");
-    currentNewsContentImgElem.src = newsCategoryImg;
-}
-
-const changeCategoryConetent = (newsCategoryIndex, clickedTabElement) => {
-    const tabTopicIdName = clickedTabElement.id;
-    const currentNewsSectionElem = document.querySelector(`section[roll="tabpanel"]`);
-
-    currentNewsSectionElem.id = `${"tabpanelTopics" + (newsCategoryIndex + 1)}`;
-    currentNewsSectionElem.setAttribute("aria-labelledby", tabTopicIdName);
-}
-
-const renderArticle = (newsContentsData) => {
-    const ul = document.createElement("ul");
-    const fragment = document.createDocumentFragment();
-
-    newsContentsData.forEach(newsArticleData => {
-        const li = document.createElement("li");
-        const h1 = document.createElement("h1");
-        const a = document.createElement("a");
-
-        const title = newsArticleData.title;
-        const commentArray = newsArticleData.comments;
-
-        a.href = "#";
-        a.textContent = title;
-
-        h1.appendChild(a)
-        li.appendChild(h1);
-
-        if (commentArray.length !== 0) {
-            li.append(createCommentIcon(commentArray));
-        }
-        fragment.appendChild(li);
-    })
-    ul.appendChild(fragment);
-
-    return ul;
-}
-
-const createCommentIcon = (commentArray) => {
-    const div = document.createElement("div");
-
-    div.classList = "commentIcon_wrap";
-    div.insertAdjacentHTML("beforeend",
-        `<span class="fa-layers fa-fw">
-            <i class="fa fa-light fa-comment faa-vertical animated"></i>
-            <span class= "fa-layers-counter" > ${commentArray.length}</span>
-        </span>`
-    );
-
-    return div;
-}
-
-const createSection = (tabTopics, index) => {
-    const section = document.createElement("section");
-    const tabTopicElem = document.getElementById(tabTopics[index].id);
-    const tabIdName = tabTopicElem.id;
-
-    section.id = `${"tabpanelTopics" + (index + 1)} `;
-    section.setAttribute("aria-labelledby", tabIdName);
-    section.setAttribute("roll", "tabpanel")
-
-    return section;
-}
-
-const createTopicImg = (imgPath) => {
-    const img = document.createElement("img");
-    const div = document.createElement("div");
-
-    div.classList = "tabTopicImg";
-    img.src = imgPath;
-    img.width = "350";
-    img.height = "250";
-
-    div.appendChild(img);
-    return div;
-}
-
-const renderLoadingImg = () => {
-    const div = document.createElement("div");
-    const img = document.createElement("img");
-
-    div.classList = "loading_wrap";
-    div.id = "js_loading_wrap"
-    img.src = "./img/loading-circle.gif";
-    img.alt = "ローディング画像"
-
-    div.appendChild(img);
-    document.body.prepend(div);
-}
-
-const hideLoadingImg = () => {
-    document.getElementById("js_loading_wrap").remove();
-}
-
-init();
